LayoutTests/media/modern-media-controls/macos-fullscreen-media-controls/macos-fullscreen-media-controls-drag-is-prevented-over-button-expected.txt

This test pauses the video, presses the fullscreen button, and drags the controls when the mouse is over a button. It should not move.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS !!mediaControls.querySelector('.buttons-container.left') became true
PASS mediaControls.querySelector('.buttons-container.left').style.width became "116px"
PASS mediaControls.querySelector('.controls-bar').style.transform is ""
PASS successfullyParsed is true

TEST COMPLETE
